+++ maldiamrkit/__init__.py
@@ -22,7 +22,7 @@ Examples
 from .dataset import MaldiSet
 from .spectrum import MaldiSpectrum
 
-__version__ = "0.11.1"
+__version__ = "0.11.2"
 __author__ = "Ettore Rocchi"
 
 __all__ = [